Fix `time_execution` function
The check for the noexceptness of the argument function has been
fixed: due to incorrect syntax the check would always evaluate to true
no matter what. Additionally, the argument function can now take any
number of arguments, which are then perfect-forwarded at call site.
Further, the chain of if's was not evaluated at compile-time in its
entirety due to the lack of `constexpr` keyword for all branches.